requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

tensorflow
streamlit
pillow
pandas
matplotlib
numpy
requests
sklearn

#Run this file with - pip3 install -r requirements.txt